+++ lib/rails_ai_bridge/server.rb
@@ -36,6 +36,7 @@ module RailsAiBridge
       Tools::GetGems,
       Tools::SearchCode,
       Tools::SearchSemantic,
+      Tools::ExplainSymbol,
       Tools::GetConventions,
       Tools::GetControllers,
       Tools::GetConfig,